refactor(org-core): split org_mode.rs into submodules
Break the 3614-line god file into focused submodules under
org-core/src/org_mode/:
  types.rs    — data types (OrgMode, TreeNode, SearchResult, etc.)
  core.rs     — file listing, search, outline, tasks, agenda query
  capture.rs  — capture_append + validation + private helpers + tests
  agenda.rs   — date utilities + AgendaViewType + tests
  tests.rs    — integration-style capture/datetree tests
  mod.rs      — module wiring and re-exports

Also moved private-function unit tests into the submodule where the
tested code lives (format_heading tests to capture.rs,
add_repeater_duration tests to agenda.rs) to avoid visibility
bumps.

Update CLAUDE.md file path references.
